src/scripts/requirement.txt
matplotlib
Pillow
numpy
pickle5
opencv-python
pandas